only make common-auth backup file if it doesn't already exist

+++ setup/_post-install/set-up-work-pop-os.sh
@@ -193,7 +193,9 @@ if [[ ! -e '/dev/sr0' ]]; then
   log 'Installing fingerprint scanner packages'
   sudo apt-get install --yes fprintd libpam-fprintd
   sudo pam-auth-update --enable fprintd
-  sudo cp '/etc/pam.d/common-auth' '/etc/pam.d/common-auth.orig'
+  if [[ ! -f '/etc/pam.d/common-auth.orig' ]]; then
+    sudo cp '/etc/pam.d/common-auth' '/etc/pam.d/common-auth.orig'
+  fi
   sudo sed --in-place "s/ max-tries=[0-9]\+ / max-tries=10 /g" '/etc/pam.d/common-auth'
   if ! fprintd-list "${USER}" | grep --quiet --fixed-strings 'right-index-finger'; then
     fprintd-enroll
